изменение вовзращаемого типа

+++ kolRand/kolRand.go
@@ -34,8 +34,8 @@ func (rnd *KolRandom) MakeUniform(n int32) uint16 {
 	return uint16(rnd.seed.Int31n(n))
 }
 
-func (rnd *KolRandom) MakeUniformFloatFromZeroToOne(n, m float32) float32 {
-	return (float32(rnd.MakeUniformRange(int32(n)*10, int32(m)*10)) / 10)
+func (rnd *KolRandom) MakeUniformFloatFromZeroToOne(n, m float32) float64 {
+	return (float64(rnd.MakeUniformRange(int32(n)*10, int32(m)*10)) / 10)
 }
 
 func (rnd *KolRandom) MakeUniformRange(n, m int32) uint16 {
